pyproject.toml
[tool.poetry]
name = "workout-ai"
version = "0.1.0"
description = "AI-powered adaptive strength & HIIT training"
authors = ["Umesh Dangat <udangat@udangat.name>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"  # or whichever version you choose
fastapi = "^0.95.0"
openai = "^1.0.0"
psycopg = "^3.1.8"  # or the current version of psycopg 3
pydantic = "^1.10.0"
alembic = "^1.8.0"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
mysql-connector-python = "^9.2.0"
faiss-cpu = "^1.10.0"
numpy = "<2"
sentence-transformers="^3.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
